# Use AddressBookCollection and regex speedup