This PR is still a bit WIP.

I am trying to use the AddressBookCollection more.  At the same time I try to sort out the different paths that can and need to .load() an AddressBook.

The "Simplify derived search regex" commit (8b98889) gives me a bit of a speed improvement.

For measurements I use [this zsh function](https://github.com/lucc/shell-config/blob/master/functions/comp) to compare the execution time of 10 runs of a command:
```
$ comp './khard-runner.py ls -f' './khard-runner.py ls -f -a other' './khard-runner.py ls -f simon' './khard-runner.py ls -f -a other simon'
1  17.33s user 0.34s system 99% cpu 17.666 total
2  6.14s user 0.27s system 100% cpu 6.394 total
3  2.48s user 0.23s system 100% cpu 2.699 total
4  1.84s user 0.16s system 100% cpu 1.994 total
```
whereas on develop I get
```
1  17.37s user 0.39s system 99% cpu 17.773 total
2  13.71s user 0.40s system 99% cpu 14.119 total
3  2.40s user 0.29s system 99% cpu 2.691 total
4  2.35s user 0.23s system 99% cpu 2.580 total
```

The differences for the "all address books" commands should be within error margin but the times for the search in only one address book looks like a real speed improvement to me.